refactor bento layout
