+++ src/components/bento/overlay/delete-button.tsx
@@ -5,7 +5,7 @@ import { Trash } from "lucide-react";
 import type * as z from "zod";
 
 import { type bentoSchema } from "@/server/db";
-import { api } from "@/trpc/client";
+import { api } from "@/trpc/react";
 import { Button } from "@/components/ui/button";
 
 export default function DeleteButton({
@@ -16,20 +16,39 @@ export default function DeleteButton({
   const router = useRouter();
   const { link } = useParams() as { link: string };
 
+  const queryClient = api.useContext();
+
+  const { mutateAsync: deleteBento } = api.profileLink.deleteBento.useMutation({
+    onMutate: () => {
+      void queryClient.profileLink.getByLink.setData(
+        {
+          link,
+        },
+        (old) => {
+          if (!old) return old;
+
+          return {
+            ...old,
+            bento: old.bento.filter((b) => b.id !== bento.id),
+          };
+        }
+      );
+    },
+    onSuccess: () => {
+      void router.refresh();
+    },
+  });
+
   return (
     <Button
       size="icon"
       variant="secondary"
       className="absolute left-0 top-0 z-20 -translate-x-1/2 -translate-y-1/2 rounded-full shadow transition-transform duration-200 ease-in-out active:scale-95"
       onClick={() => {
-        void api.profileLink.deleteBento
-          .mutate({
-            link,
-            id: bento.id,
-          })
-          .then(() => {
-            void router.refresh();
-          });
+        void deleteBento({
+          link,
+          id: bento.id,
+        });
       }}
     >
       <Trash className="h-[1.2rem] w-[1.2rem]" />
